new file systems works

+++ libvips/file/jpeg.h
@@ -53,7 +53,7 @@ void vips__new_error_exit( j_common_ptr cinfo );
 int vips__jpeg_write_file( VipsImage *in, 
 	const char *filename, int Q, const char *profile );
 int vips__jpeg_write_buffer( VipsImage *in, 
-	void **obuf, int *olen, int Q, const char *profile );
+	void **obuf, size_t *olen, int Q, const char *profile );
 
 int vips__isjpeg( const char *filename );
 int vips__jpeg_read_file( const char *name, VipsImage *out,
